.tap that path!
The install.rb script is missing .tap on the importmap.path and so does actually set any of the passed assets/directories. The tests did not pick up on this,  because the dummy app correctly uses .tap in it's assets.rb initializer. It's just the one created by the install script that has issues :)